requirements.txt
hyperliquid-python-sdk>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
eth-account>=0.11.0
schedule>=1.2.0
flask>=3.0.0
discord.py>=2.3.0
